@@ -1,6 +1,6 @@
 package io.ipinfo.spring.strategies.interceptor;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 public interface InterceptorStrategy{
     boolean shouldRun(HttpServletRequest request);
